+++ src/coq_elpi_builtins.ml
@@ -3884,6 +3884,17 @@ Supported attributes:
      state, !: s, [])),
   DocAbove);
 
+  LPDoc "-- Extra Dependencies -----------------------------------------------";
+
+  MLCode(Pred("coq.extra-dep",
+    In(id, "Identifier",
+    Out(option id, "File Name",
+    Easy "Resolve the file name of an Extra Dependency")),
+  (fun id _ ~depth ->
+    !: (try Some (ComExtraDeps.query_extra_dep (Names.Id.of_string id))
+        with Not_found -> None))),
+  DocAbove);
+
   LPDoc "-- Access to Elpi's data --------------------------------------------";
 
    (* Self modification *)